+++ components/firstpair/impl/src/main/java/com/flipperdevices/firstpair/impl/composable/searching/ComposableSearchingDevices.kt
@@ -55,7 +55,7 @@ fun ComposableSearchingDevices(
         onRefresh = onRefreshSearching
     ) {
         LazyColumn(
-            modifier = it.padding(vertical = 18.dp),
+            modifier = Modifier.padding(vertical = 18.dp),
             verticalArrangement = Arrangement.spacedBy(10.dp)
         ) {
             items(